Move game management into server
